+++ src/main/java/roomescape/reservation/application/usecase/ReservationQueryUseCase.java
@@ -1,7 +1,7 @@
 package roomescape.reservation.application.usecase;
 
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation_time.domain.ReservationTimeId;
+import roomescape.time.domain.ReservationTimeId;
 
 import java.time.LocalDate;
 import java.util.List;
