--- src/main/java/roomescape/reservation_time/application/ReservationTimeServiceImpl.java
@@ -1,40 +1,0 @@
-package roomescape.reservation_time.application;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import roomescape.reservation_time.application.converter.ReservationTimeConverter;
-import roomescape.reservation_time.application.dto.CreateReservationTimeServiceRequest;
-import roomescape.reservation_time.application.usecase.ReservationTimeCommandUseCase;
-import roomescape.reservation_time.application.usecase.ReservationTimeQueryUseCase;
-import roomescape.reservation_time.domain.ReservationTimeId;
-import roomescape.reservation_time.ui.dto.CreateReservationTimeWebRequest;
-import roomescape.reservation_time.ui.dto.ReservationTimeResponse;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class ReservationTimeServiceImpl implements ReservationTimeService {
-
-    private final ReservationTimeQueryUseCase reservationTimeQueryUseCase;
-    private final ReservationTimeCommandUseCase reservationTimeCommandUseCase;
-
-    @Override
-    public List<ReservationTimeResponse> getAll() {
-        return ReservationTimeConverter.toDto(
-                reservationTimeQueryUseCase.getAll());
-    }
-
-    @Override
-    public ReservationTimeResponse create(final CreateReservationTimeWebRequest createReservationTimeWebRequest) {
-        return ReservationTimeConverter.toDto(
-                reservationTimeCommandUseCase.create(
-                        new CreateReservationTimeServiceRequest(
-                                createReservationTimeWebRequest.startAt())));
-    }
-
-    @Override
-    public void delete(final ReservationTimeId id) {
-        reservationTimeCommandUseCase.delete(id);
-    }
-}
