--- src/main/java/roomescape/reservation_time/application/converter/ReservationTimeConverter.java
@@ -1,36 +1,0 @@
-package roomescape.reservation_time.application.converter;
-
-import roomescape.reservation_time.application.dto.CreateReservationTimeServiceRequest;
-import roomescape.reservation_time.domain.ReservationTime;
-import roomescape.reservation_time.domain.ReservationTimeId;
-import roomescape.reservation_time.infrastructure.entity.ReservationTimeEntity;
-import roomescape.reservation_time.ui.dto.ReservationTimeResponse;
-
-import java.util.List;
-
-public class ReservationTimeConverter {
-
-    public static ReservationTime toDomain(final CreateReservationTimeServiceRequest request) {
-        return ReservationTime.of(
-                ReservationTimeId.unassigned(),
-                request.startAt());
-    }
-
-    public static ReservationTime toDomain(final ReservationTimeEntity entity) {
-        return ReservationTime.of(
-                ReservationTimeId.from(entity.getId()),
-                entity.getTime().toLocalTime());
-    }
-
-    public static ReservationTimeResponse toDto(final ReservationTime reservationTime) {
-        return new ReservationTimeResponse(
-                reservationTime.getId().getValue(),
-                reservationTime.getValue());
-    }
-
-    public static List<ReservationTimeResponse> toDto(final List<ReservationTime> reservationTimes) {
-        return reservationTimes.stream()
-                .map(ReservationTimeConverter::toDto)
-                .toList();
-    }
-}
