--- src/main/java/roomescape/reservation_time/ui/dto/CreateReservationTimeWebRequest.java
@@ -1,20 +1,0 @@
-package roomescape.reservation_time.ui.dto;
-
-import lombok.AccessLevel;
-import lombok.experimental.FieldNameConstants;
-import roomescape.common.validate.Validator;
-
-import java.time.LocalTime;
-
-@FieldNameConstants(level = AccessLevel.PRIVATE)
-public record CreateReservationTimeWebRequest(LocalTime startAt) {
-
-    public CreateReservationTimeWebRequest {
-        validate(startAt);
-    }
-
-    private void validate(final LocalTime startAt) {
-        Validator.of(CreateReservationTimeWebRequest.class)
-                .notNullField(Fields.startAt, startAt);
-    }
-}
