+++ src/main/java/roomescape/time/application/converter/ReservationTimeConverter.java
@@ -1,0 +1,36 @@
+package roomescape.time.application.converter;
+
+import roomescape.time.application.dto.CreateReservationTimeServiceRequest;
+import roomescape.time.domain.ReservationTime;
+import roomescape.time.domain.ReservationTimeId;
+import roomescape.time.infrastructure.entity.ReservationTimeEntity;
+import roomescape.time.ui.dto.ReservationTimeResponse;
+
+import java.util.List;
+
+public class ReservationTimeConverter {
+
+    public static ReservationTime toDomain(final CreateReservationTimeServiceRequest request) {
+        return ReservationTime.of(
+                ReservationTimeId.unassigned(),
+                request.startAt());
+    }
+
+    public static ReservationTime toDomain(final ReservationTimeEntity entity) {
+        return ReservationTime.of(
+                ReservationTimeId.from(entity.getId()),
+                entity.getTime().toLocalTime());
+    }
+
+    public static ReservationTimeResponse toDto(final ReservationTime reservationTime) {
+        return new ReservationTimeResponse(
+                reservationTime.getId().getValue(),
+                reservationTime.getValue());
+    }
+
+    public static List<ReservationTimeResponse> toDto(final List<ReservationTime> reservationTimes) {
+        return reservationTimes.stream()
+                .map(ReservationTimeConverter::toDto)
+                .toList();
+    }
+}
