+++ src/main/java/roomescape/time/infrastructure/entity/ReservationTimeEntity.java
@@ -1,0 +1,19 @@
+package roomescape.time.infrastructure.entity;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.sql.Time;
+
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class ReservationTimeEntity {
+
+    private final Long id;
+    private final Time time;
+
+    public static ReservationTimeEntity of(final Long id, final Time time) {
+        return new ReservationTimeEntity(id, time);
+    }
+}
