+++ src/test/java/roomescape/reservation/domain/ReservationTest.java
@@ -3,8 +3,8 @@ package roomescape.reservation.domain;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.reservation_time.domain.ReservationTime;
-import roomescape.reservation_time.domain.ReservationTimeId;
+import roomescape.time.domain.ReservationTime;
+import roomescape.time.domain.ReservationTimeId;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
